Écris un premier test du module public paramètres
