+++ test_public/modules/parametres.spec.mjs
@@ -1,0 +1,19 @@
+import expect from 'expect.js';
+import parametres from '../../public/modules/parametres.mjs';
+import { JSDOM } from 'jsdom';
+import jquery from 'jquery';
+
+describe('les paramètres', () => {
+  before(() => {
+    const sourcePage = `<form class="formulaire"><input name="champs-1" value="valeur 1"></form>`;
+    const dom = new JSDOM(sourcePage);
+    global.$ = jquery(dom.window);
+  });
+  
+  it('renvoient les données du formulaire', () => {
+    const parametresFormulaire = parametres('.formulaire');
+
+    expect(parametresFormulaire).to.have.key('champs-1');
+    expect(parametresFormulaire['champs-1']).to.be('valeur 1');
+  });
+});
